requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.6.0
certifi==2022.12.7
click==8.0.4
Django==4.1.7
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
pip==23.0.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.6
pytest==7.2.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2022.7.1
setuptools==65.6.3
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
wheel==0.38.4
